refactor(torghut): split runtime window importer ownership
