+++ services/torghut/scripts/import_hypothesis_runtime_windows.py
@@ -221,6 +221,65 @@ from scripts.hypothesis_runtime_window_import.source_windows import (
     _order_lifecycle_query_row_has_time as _order_lifecycle_query_row_has_time,
 )
 
+from scripts.hypothesis_runtime_window_import.execution_costs import (
+    _runtime_execution_cost_amount as _runtime_execution_cost_amount,
+    _runtime_execution_cost_basis as _runtime_execution_cost_basis,
+    _strategy_name_candidates as _strategy_name_candidates,
+)
+from scripts.hypothesis_runtime_window_import.lifecycle_rows import (
+    _execution_id_by_order_id as _execution_id_by_order_id,
+    _execution_side_by_order_id as _execution_side_by_order_id,
+    _fill_quantity_basis as _fill_quantity_basis,
+    _order_feed_fill_delta_blockers as _order_feed_fill_delta_blockers,
+    _order_feed_fill_lifecycle_blockers as _order_feed_fill_lifecycle_blockers,
+    _runtime_lifecycle_identifier_values as _runtime_lifecycle_identifier_values,
+    _runtime_lifecycle_ledger_row as _runtime_lifecycle_ledger_row,
+    _runtime_lifecycle_strategy_identifier_values as _runtime_lifecycle_strategy_identifier_values,
+    _runtime_order_id as _runtime_order_id,
+    _source_order_feed_payload_delta_fill as _source_order_feed_payload_delta_fill,
+    _strategy_relevant_unlinked_fill_lifecycle_rows as _strategy_relevant_unlinked_fill_lifecycle_rows,
+    _with_linked_execution_id as _with_linked_execution_id,
+)
+from scripts.hypothesis_runtime_window_import.source_row_filters import (
+    _active_carry_in_ledger_rows as _active_carry_in_ledger_rows,
+    _adjust_carry_in_lot_row as _adjust_carry_in_lot_row,
+    _event_sourced_fill_economics_order_ids as _event_sourced_fill_economics_order_ids,
+    _execution_fill_economics_order_ids as _execution_fill_economics_order_ids,
+    _execution_tca_order_ids as _execution_tca_order_ids,
+    _filter_carry_in_source_rows_for_active_lots as _filter_carry_in_source_rows_for_active_lots,
+    _flat_start_position_snapshot_from_cursor as _flat_start_position_snapshot_from_cursor,
+    _ledger_row_decimal as _ledger_row_decimal,
+    _ledger_row_decision_id as _ledger_row_decision_id,
+    _ledger_row_event_type as _ledger_row_event_type,
+    _ledger_row_order_id as _ledger_row_order_id,
+    _ledger_row_position_key as _ledger_row_position_key,
+    _ledger_row_side_sign as _ledger_row_side_sign,
+    _ledger_row_symbol as _ledger_row_symbol,
+    _ledger_row_text as _ledger_row_text,
+    _ledger_row_time as _ledger_row_time,
+    _ledger_row_value as _ledger_row_value,
+    _required_order_lifecycle_source_row_count as _required_order_lifecycle_source_row_count,
+    _runtime_decision_rows_before_bucket as _runtime_decision_rows_before_bucket,
+    _runtime_decision_rows_for_bucket as _runtime_decision_rows_for_bucket,
+    _runtime_order_rows_for_bucket as _runtime_order_rows_for_bucket,
+    _runtime_source_decision_ids as _runtime_source_decision_ids,
+    _runtime_source_row_before_bucket as _runtime_source_row_before_bucket,
+    _runtime_source_row_in_bucket as _runtime_source_row_in_bucket,
+    _runtime_unlinked_order_rows_for_bucket as _runtime_unlinked_order_rows_for_bucket,
+    _source_backed_fill_lifecycle_order_ids as _source_backed_fill_lifecycle_order_ids,
+    _source_backed_fill_lifecycle_rows as _source_backed_fill_lifecycle_rows,
+    _source_backed_order_lifecycle_rows as _source_backed_order_lifecycle_rows,
+    _source_backed_submitted_lifecycle_order_ids as _source_backed_submitted_lifecycle_order_ids,
+)
+from scripts.hypothesis_runtime_window_import.source_activity_context import (
+    _runtime_execution_ledger_fill_from_row as _runtime_execution_ledger_fill_from_row,
+    _runtime_source_context_for_bucket as _runtime_source_context_for_bucket,
+    _source_activity_diagnostics_blockers as _source_activity_diagnostics_blockers,
+)
+from scripts.hypothesis_runtime_window_import.realized_pnl import (
+    _build_realized_strategy_pnl_rows as _build_realized_strategy_pnl_rows,
+)
+
 
 def _parse_args() -> argparse.Namespace:
     parser = argparse.ArgumentParser(
@@ -329,2393 +388,6 @@ def _persistence_session(args: argparse.Namespace) -> Iterator[Session]:
 
 def _flag(value: str) -> bool:
     return value.strip().lower() not in {"", "0", "false", "no", "off"}
-
-
-def _strategy_name_candidates(*values: str | None) -> list[str]:
-    candidates: list[str] = []
-    for value in values:
-        raw = str(value or "").strip()
-        if not raw:
-            continue
-        variants = [
-            raw,
-            raw.split("@", 1)[0],
-            raw.replace("_", "-"),
-            raw.split("@", 1)[0].replace("_", "-"),
-        ]
-        for variant in variants:
-            normalized = variant.strip()
-            if normalized and normalized not in candidates:
-                candidates.append(normalized)
-    return candidates
-
-
-def _runtime_execution_cost_amount(
-    row: Mapping[str, object],
-    *,
-    filled_notional: Decimal,
-    side: Any = None,
-    filled_qty: Decimal | None = None,
-) -> Decimal | None:
-    explicit_cost = _first_decimal(
-        row,
-        "cost_amount",
-        "explicit_cost",
-        "commission",
-        "fees",
-        "fee_amount",
-        "broker_fee",
-    )
-    if explicit_cost is not None:
-        return explicit_cost
-    fee_schedule_cost = _alpaca_2026_equity_fee_schedule_cost(
-        row,
-        side=side,
-        filled_qty=filled_qty,
-        filled_notional=filled_notional,
-    )
-    if fee_schedule_cost is not None:
-        return fee_schedule_cost[0]
-    total_cost_bps = _first_decimal(
-        row,
-        "total_cost_bps",
-        "estimated_total_cost_bps",
-        "explicit_cost_bps",
-        "execution_total_cost_bps",
-    )
-    if total_cost_bps is None or total_cost_bps < 0 or filled_notional <= 0:
-        return None
-    return filled_notional * total_cost_bps / Decimal("10000")
-
-
-def _runtime_execution_cost_basis(
-    row: Mapping[str, object],
-    *,
-    cost_amount: Decimal | None,
-    side: Any = None,
-    filled_qty: Decimal | None = None,
-    filled_notional: Decimal | None = None,
-) -> str | None:
-    cost_basis = _first_text(
-        row,
-        "cost_basis",
-        "cost_source",
-        "fee_basis",
-        "commission_basis",
-        "broker_fee_basis",
-    )
-    if cost_basis is not None:
-        return cost_basis
-    if filled_notional is not None:
-        fee_schedule_cost = _alpaca_2026_equity_fee_schedule_cost(
-            row,
-            side=side,
-            filled_qty=filled_qty,
-            filled_notional=filled_notional,
-        )
-        if fee_schedule_cost is not None and cost_amount == fee_schedule_cost[0]:
-            return fee_schedule_cost[1]
-    if (
-        cost_amount is not None
-        and _first_decimal(
-            row,
-            "total_cost_bps",
-            "estimated_total_cost_bps",
-            "explicit_cost_bps",
-            "execution_total_cost_bps",
-        )
-        is not None
-    ):
-        return "decision_impact_assumptions_total_cost_bps"
-    return None
-
-
-def _runtime_lifecycle_ledger_row(
-    row: Mapping[str, object],
-    *,
-    event_type: str,
-    require_complete_fill: bool = False,
-) -> dict[str, object] | None:
-    event_time = _runtime_ledger_row_time(
-        {str(key): value for key, value in row.items()}
-    )
-    if event_time is None:
-        return None
-    symbol = _first_text(row, "symbol", "order_symbol")
-    ledger_row: dict[str, object] = {
-        "executed_at": event_time,
-        "event_type": event_type,
-        "account_label": _first_text(row, "account_label", "alpaca_account_label"),
-        "strategy_id": _first_text(row, "strategy_id", "strategy_name"),
-        "symbol": symbol.strip().upper() if symbol is not None else None,
-        "decision_id": _first_text(
-            row,
-            "decision_id",
-            "trade_decision_id",
-            "decision_hash",
-        ),
-        "order_id": _first_text(
-            row,
-            "order_id",
-            "alpaca_order_id",
-            "client_order_id",
-            "execution_correlation_id",
-        ),
-        "execution_policy_hash": _first_text(
-            row,
-            "execution_policy_hash",
-            "execution_policy_sha256",
-            "policy_hash",
-        )
-        or _first_payload_digest(
-            row,
-            "execution_policy",
-            "execution_policy_context",
-            "execution_advisor",
-            "_execution_advice_provenance",
-        ),
-        "cost_model_hash": _first_text(
-            row,
-            "cost_model_hash",
-            "fee_model_hash",
-            "cost_model_sha256",
-        )
-        or _first_payload_digest(
-            row,
-            "cost_model",
-            "cost_model_config",
-            "transaction_cost_model",
-            "fee_model",
-            "fees_model",
-        ),
-        "lineage_hash": _first_text(
-            row,
-            "lineage_hash",
-            "candidate_lineage_hash",
-            "replay_lineage_hash",
-            "candidate_evaluation_key",
-            "event_fingerprint",
-        )
-        or _first_lineage_digest(row),
-        "replay_data_hash": _first_text(
-            row,
-            "replay_data_hash",
-            "replay_tape_content_sha256",
-            "dataset_snapshot_hash",
-            "source_query_digest",
-            "source_offset",
-        ),
-    }
-    if _source_authority_order_event_row(row):
-        ledger_row["source"] = "execution_order_event"
-        ledger_row["source_materialization"] = "execution_order_events"
-        ledger_row["authority_class"] = "runtime_order_feed_execution_source"
-        for key in (
-            "execution_order_event_id",
-            "execution_id",
-            "source_topic",
-            "source_partition",
-            "source_offset",
-            "source_window_id",
-        ):
-            if row.get(key) is not None:
-                ledger_row[key] = row[key]
-    if event_type in _RUNTIME_LEDGER_FILL_EVENTS:
-        side = _first_text(row, "side", "action", "order_side")
-        source_authority_order_event = _source_authority_order_event_row(row)
-        fill_quantity_basis = _fill_quantity_basis(row)
-        filled_qty_delta = _first_positive_decimal(
-            row,
-            "filled_qty_delta",
-            "fill_qty_delta",
-            "delta_filled_qty",
-        )
-        filled_qty = _first_positive_decimal(
-            row,
-            "filled_qty",
-            "filled_quantity",
-            "qty",
-            "quantity",
-        )
-        avg_fill_price = _first_positive_decimal(
-            row,
-            "avg_fill_price",
-            "filled_avg_price",
-            "filled_average_price",
-            "average_fill_price",
-            "fill_price",
-            "filled_price",
-            "price",
-        )
-        source_order_feed_delta_fill = (
-            _source_order_feed_payload_delta_fill(row, event_type=event_type)
-            if source_authority_order_event and fill_quantity_basis is None
-            else None
-        )
-        if (
-            source_authority_order_event
-            and fill_quantity_basis in {"delta", "cumulative_to_delta"}
-            and filled_qty_delta is not None
-        ):
-            filled_qty = filled_qty_delta
-        filled_notional_delta = _first_positive_decimal(
-            row,
-            "filled_notional_delta",
-            "fill_notional_delta",
-            "delta_filled_notional",
-        )
-        filled_notional = _first_positive_decimal(
-            row,
-            "filled_notional_delta",
-            "fill_notional_delta",
-            "delta_filled_notional",
-            "filled_notional",
-            "notional",
-            "fill_notional",
-        )
-        if source_order_feed_delta_fill is not None:
-            delta_qty, delta_price, delta_side = source_order_feed_delta_fill
-            side = side or delta_side
-            fill_quantity_basis = "delta"
-            filled_qty_delta = delta_qty
-            filled_qty = delta_qty
-            avg_fill_price = delta_price
-            filled_notional_delta = delta_qty * delta_price
-            filled_notional = filled_notional_delta
-        elif source_authority_order_event and fill_quantity_basis not in {
-            "delta",
-            "cumulative_to_delta",
-        }:
-            filled_qty = None
-            filled_notional = None
-        elif (
-            source_authority_order_event
-            and fill_quantity_basis in {"delta", "cumulative_to_delta"}
-            and filled_qty_delta is None
-        ):
-            filled_qty = None
-            filled_notional = None
-        if (
-            filled_notional is None
-            and filled_qty is not None
-            and avg_fill_price is not None
-        ):
-            filled_notional = filled_qty * avg_fill_price
-        cost_amount = (
-            _runtime_execution_cost_amount(
-                row,
-                filled_notional=filled_notional,
-                side=side,
-                filled_qty=filled_qty,
-            )
-            if filled_notional is not None
-            else None
-        )
-        cost_basis = _runtime_execution_cost_basis(
-            row,
-            cost_amount=cost_amount,
-            side=side,
-            filled_qty=filled_qty,
-            filled_notional=filled_notional,
-        )
-        if require_complete_fill and (
-            side is None
-            or filled_qty is None
-            or filled_qty <= 0
-            or avg_fill_price is None
-            or avg_fill_price <= 0
-            or filled_notional is None
-            or filled_notional <= 0
-            or cost_amount is None
-            or cost_amount < 0
-            or cost_basis is None
-        ):
-            return None
-        if side is not None:
-            ledger_row["side"] = side
-        if filled_qty is not None:
-            ledger_row["filled_qty"] = filled_qty
-        if avg_fill_price is not None:
-            ledger_row["avg_fill_price"] = avg_fill_price
-        if filled_notional is not None:
-            ledger_row["filled_notional"] = filled_notional
-        if filled_qty_delta is not None:
-            ledger_row["filled_qty_delta"] = filled_qty_delta
-        if filled_notional_delta is not None:
-            ledger_row["filled_notional_delta"] = filled_notional_delta
-        if fill_quantity_basis is not None:
-            ledger_row["fill_quantity_basis"] = fill_quantity_basis
-        if cost_amount is not None:
-            ledger_row["cost_amount"] = cost_amount
-        if cost_basis is not None:
-            ledger_row["cost_basis"] = cost_basis
-        if _cost_basis_is_alpaca_fee_schedule(cost_basis):
-            ledger_row["cost_model_hash"] = _alpaca_2026_equity_fee_schedule_hash()
-    return ledger_row
-
-
-def _runtime_order_id(row: Mapping[str, object]) -> str | None:
-    return _first_text(
-        row,
-        "order_id",
-        "alpaca_order_id",
-        "client_order_id",
-        "execution_correlation_id",
-    )
-
-
-def _execution_id_by_order_id(
-    rows: Sequence[Mapping[str, object]],
-) -> dict[str, str]:
-    ids_by_order_id: dict[str, set[str]] = {}
-    for row in rows:
-        order_id = _runtime_order_id(row)
-        execution_id = _first_text(row, "execution_id")
-        if order_id is None or execution_id is None:
-            continue
-        ids_by_order_id.setdefault(order_id, set()).add(execution_id)
-    return {
-        order_id: next(iter(execution_ids))
-        for order_id, execution_ids in ids_by_order_id.items()
-        if len(execution_ids) == 1
-    }
-
-
-def _execution_side_by_order_id(
-    rows: Sequence[Mapping[str, object]],
-) -> dict[str, str]:
-    sides_by_order_id: dict[str, set[str]] = {}
-    for row in rows:
-        order_id = _runtime_order_id(row)
-        side = _first_text(row, "side", "order_side")
-        if order_id is None or side is None:
-            continue
-        sides_by_order_id.setdefault(order_id, set()).add(side)
-    return {
-        order_id: next(iter(sides))
-        for order_id, sides in sides_by_order_id.items()
-        if len(sides) == 1
-    }
-
-
-def _with_linked_execution_id(
-    row: Mapping[str, object],
-    *,
-    execution_id_by_order_id: Mapping[str, str],
-    execution_side_by_order_id: Mapping[str, str] | None = None,
-) -> dict[str, object]:
-    payload = dict(row)
-    order_id = _runtime_order_id(payload)
-    if (
-        _first_text(payload, "execution_id") is None
-        and order_id is not None
-        and (execution_id := execution_id_by_order_id.get(order_id))
-    ):
-        payload["execution_id"] = execution_id
-    if (
-        _first_text(payload, "side", "order_side") is None
-        and order_id is not None
-        and execution_side_by_order_id is not None
-        and (side := execution_side_by_order_id.get(order_id))
-    ):
-        payload["side"] = side
-    return payload
-
-
-def _runtime_lifecycle_identifier_values(row: Mapping[str, object]) -> set[str]:
-    values = _text_values(row, *_RUNTIME_LIFECYCLE_IDENTIFIER_KEYS)
-    return {value for value in values if value}
-
-
-def _runtime_lifecycle_strategy_identifier_values(
-    rows: Sequence[Mapping[str, object]],
-) -> set[str]:
-    identifiers: set[str] = set()
-    for row in rows:
-        identifiers.update(_runtime_lifecycle_identifier_values(row))
-    return identifiers
-
-
-def _strategy_relevant_unlinked_fill_lifecycle_rows(
-    *,
-    execution_rows: Sequence[Mapping[str, object]],
-    order_lifecycle_rows: Sequence[Mapping[str, object]] | None,
-    unlinked_order_lifecycle_rows: Sequence[Mapping[str, object]] | None,
-    candidate_id: str | None = None,
-    hypothesis_id: str | None = None,
-) -> list[dict[str, object]]:
-    strategy_identifiers = _runtime_lifecycle_strategy_identifier_values(
-        [
-            *execution_rows,
-            *(order_lifecycle_rows or ()),
-        ]
-    )
-    lineage_filter_present = candidate_id is not None or hypothesis_id is not None
-    relevant_rows: list[dict[str, object]] = []
-    for row in unlinked_order_lifecycle_rows or ():
-        normalized = {str(key): value for key, value in row.items()}
-        if _runtime_ledger_event_type(normalized) not in _RUNTIME_LEDGER_FILL_EVENTS:
-            continue
-        if lineage_filter_present and _source_row_matches_lineage(
-            normalized,
-            candidate_id=candidate_id,
-            hypothesis_id=hypothesis_id,
-            require_source_lineage=True,
-        ):
-            relevant_rows.append(normalized)
-            continue
-        if (
-            strategy_identifiers
-            and _runtime_lifecycle_identifier_values(normalized) & strategy_identifiers
-        ):
-            relevant_rows.append(normalized)
-    return relevant_rows
-
-
-def _order_feed_fill_lifecycle_blockers(
-    *,
-    execution_rows: list[dict[str, object]],
-    order_lifecycle_rows: list[dict[str, object]] | None,
-    unlinked_order_lifecycle_rows: list[dict[str, object]] | None = None,
-) -> list[str]:
-    expected_order_ids: set[str] = set()
-    missing_execution_order_id = False
-    for row in execution_rows:
-        if not _execution_row_has_fill(row):
-            continue
-        order_id = _runtime_order_id(row)
-        if order_id is None:
-            missing_execution_order_id = True
-            continue
-        expected_order_ids.add(order_id)
-
-    if not expected_order_ids and not missing_execution_order_id:
-        return []
-
-    observed_fill_order_ids: set[str] = set()
-    for row in order_lifecycle_rows or []:
-        normalized = {str(key): value for key, value in row.items()}
-        event_type = _runtime_ledger_event_type(normalized)
-        if event_type not in _RUNTIME_LEDGER_FILL_EVENTS:
-            continue
-        order_id = _runtime_order_id(row)
-        if order_id is None:
-            continue
-        observed_fill_order_ids.add(order_id)
-
-    blockers: list[str] = []
-    if _strategy_relevant_unlinked_fill_lifecycle_rows(
-        execution_rows=execution_rows,
-        order_lifecycle_rows=order_lifecycle_rows,
-        unlinked_order_lifecycle_rows=unlinked_order_lifecycle_rows,
-    ):
-        blockers.append(ORDER_FEED_UNLINKED_FILL_LIFECYCLE_PRESENT_BLOCKER)
-    if missing_execution_order_id:
-        blockers.append(ORDER_FEED_FILL_LIFECYCLE_ORDER_ID_MISSING_BLOCKER)
-    if not observed_fill_order_ids:
-        blockers.append(ORDER_FEED_FILL_LIFECYCLE_MISSING_BLOCKER)
-    elif expected_order_ids - observed_fill_order_ids:
-        blockers.append(ORDER_FEED_FILL_LIFECYCLE_INCOMPLETE_BLOCKER)
-    return blockers
-
-
-def _fill_quantity_basis(row: Mapping[str, object]) -> str | None:
-    raw = _first_text(row, "fill_quantity_basis", "quantity_basis")
-    if raw is None:
-        return None
-    normalized = raw.strip().lower().replace("-", "_").replace(" ", "_")
-    if normalized in {"delta", "fill_delta", "filled_delta"}:
-        return "delta"
-    if normalized in {"cumulative_to_delta", "cumulative_delta", "cum_to_delta"}:
-        return "cumulative_to_delta"
-    if normalized in {"cumulative", "cum"}:
-        return "cumulative"
-    if normalized in {"unknown", "cumulative_non_increasing"}:
-        return normalized
-    return normalized or None
-
-
-def _source_order_feed_payload_delta_fill(
-    row: Mapping[str, object], *, event_type: str
-) -> tuple[Decimal, Decimal, str | None] | None:
-    if (
-        _runtime_ledger_event_type({"event_type": event_type})
-        not in _RUNTIME_LEDGER_FILL_EVENTS
-    ):
-        return None
-    raw_event = _as_mapping(row.get("raw_event"))
-    if not raw_event:
-        return None
-
-    payload_candidates: list[Mapping[str, object]] = []
-    for key in ("payload", "data"):
-        payload = _as_mapping(raw_event.get(key))
-        if payload:
-            payload_candidates.append(payload)
-    if _as_mapping(raw_event.get("order")):
-        payload_candidates.append(raw_event)
-
-    for payload in payload_candidates:
-        payload_event = _runtime_ledger_event_type(
-            {
-                "event_type": _direct_text(payload, "event", "event_type")
-                or event_type,
-                "status": _direct_text(payload, "status"),
-            }
-        )
-        if payload_event not in _RUNTIME_LEDGER_FILL_EVENTS:
-            continue
-        qty = _decimal_or_none(payload.get("qty"))
-        price = _decimal_or_none(payload.get("price"))
-        if qty is None or qty <= 0 or price is None or price <= 0:
-            continue
-        order = _as_mapping(payload.get("order"))
-        side = _direct_text(payload, "side", "order_side") or _direct_text(
-            order, "side"
-        )
-        return qty, price, side
-    return None
-
-
-def _order_feed_fill_delta_blockers(row: Mapping[str, object]) -> list[str]:
-    if not _source_authority_order_event_row(row):
-        return []
-    if _source_order_feed_payload_delta_fill(
-        row, event_type=_runtime_ledger_event_type(row)
-    ):
-        return []
-    basis = _fill_quantity_basis(row)
-    if basis not in {"delta", "cumulative_to_delta"}:
-        return [ORDER_FEED_FILL_DELTA_BASIS_MISSING_BLOCKER]
-    if (
-        _first_positive_decimal(
-            row,
-            "filled_qty_delta",
-            "fill_qty_delta",
-            "delta_filled_qty",
-        )
-        is None
-    ):
-        return [ORDER_FEED_FILL_DELTA_MISSING_BLOCKER]
-    return []
-
-
-def _runtime_source_row_in_bucket(
-    row: Mapping[str, object],
-    *,
-    bucket: RuntimeLedgerBucket,
-) -> bool:
-    row_symbol = _runtime_source_row_symbol(row)
-    if (
-        bucket.symbol is not None
-        and row_symbol is not None
-        and row_symbol != bucket.symbol
-    ):
-        return False
-    event_time = _runtime_ledger_row_time(
-        {str(key): value for key, value in row.items()}
-    )
-    return (
-        event_time is None
-        or bucket.bucket_started_at <= event_time < bucket.bucket_ended_at
-    )
-
-
-def _runtime_source_row_before_bucket(
-    row: Mapping[str, object],
-    *,
-    bucket: RuntimeLedgerBucket,
-) -> bool:
-    row_symbol = _runtime_source_row_symbol(row)
-    if (
-        bucket.symbol is not None
-        and row_symbol is not None
-        and row_symbol != bucket.symbol
-    ):
-        return False
-    event_time = _runtime_ledger_row_time(
-        {str(key): value for key, value in row.items()}
-    )
-    return event_time is not None and event_time < bucket.bucket_started_at
-
-
-def _runtime_source_decision_ids(rows: Sequence[Mapping[str, object]]) -> set[str]:
-    identifiers: set[str] = set()
-    for row in rows:
-        for key in ("trade_decision_id", "decision_id", "decision_hash"):
-            value = _text_or_none(row.get(key))
-            if value is not None:
-                identifiers.add(value)
-    return identifiers
-
-
-def _ledger_row_value(
-    row: RuntimeLedgerFill | Mapping[str, object],
-    *keys: str,
-) -> object | None:
-    if isinstance(row, Mapping):
-        for key in keys:
-            if key in row:
-                return row.get(key)
-        return None
-    for key in keys:
-        value = getattr(row, key, None)
-        if value is not None:
-            return value
-    return None
-
-
-def _ledger_row_text(
-    row: RuntimeLedgerFill | Mapping[str, object],
-    *keys: str,
-) -> str | None:
-    for key in keys:
-        value = _ledger_row_value(row, key)
-        if (text := _text_or_none(value)) is not None:
-            return text
-    return None
-
-
-def _ledger_row_decimal(
-    row: RuntimeLedgerFill | Mapping[str, object],
-    *keys: str,
-) -> Decimal | None:
-    for key in keys:
-        value = _ledger_row_value(row, key)
-        if (parsed := _decimal_or_none(value)) is not None:
-            return parsed
-    return None
-
-
-def _ledger_row_time(row: RuntimeLedgerFill | Mapping[str, object]) -> datetime | None:
-    if isinstance(row, Mapping):
-        return _runtime_ledger_row_time({str(key): value for key, value in row.items()})
-    value = getattr(row, "executed_at", None)
-    return value if isinstance(value, datetime) else None
-
-
-def _ledger_row_event_type(row: RuntimeLedgerFill | Mapping[str, object]) -> str:
-    if isinstance(row, Mapping):
-        return _runtime_ledger_event_type(
-            {str(key): value for key, value in row.items()}
-        )
-    return str(getattr(row, "event_type", "") or "").strip().lower()
-
-
-def _ledger_row_order_id(row: RuntimeLedgerFill | Mapping[str, object]) -> str | None:
-    return _ledger_row_text(
-        row,
-        "order_id",
-        "alpaca_order_id",
-        "client_order_id",
-        "execution_correlation_id",
-    )
-
-
-def _ledger_row_decision_id(
-    row: RuntimeLedgerFill | Mapping[str, object],
-) -> str | None:
-    return _ledger_row_text(row, "decision_id", "trade_decision_id", "decision_hash")
-
-
-def _ledger_row_symbol(row: RuntimeLedgerFill | Mapping[str, object]) -> str | None:
-    symbol = _ledger_row_text(row, "symbol", "ticker")
-    return symbol.upper() if symbol is not None else None
-
-
-def _ledger_row_position_key(
-    row: RuntimeLedgerFill | Mapping[str, object],
-) -> tuple[str | None, str | None, str | None]:
-    return (
-        _ledger_row_text(row, "account_label", "alpaca_account_label"),
-        _ledger_row_text(row, "strategy_id", "strategy_name"),
-        _ledger_row_symbol(row),
-    )
-
-
-def _ledger_row_side_sign(row: RuntimeLedgerFill | Mapping[str, object]) -> Decimal:
-    side = (_ledger_row_text(row, "side", "order_side") or "").strip().lower()
-    if side in {"buy", "long", "b"}:
-        return Decimal("1")
-    if side in {"sell", "short", "s"}:
-        return Decimal("-1")
-    return Decimal("0")
-
-
-def _adjust_carry_in_lot_row(
-    row: RuntimeLedgerFill | Mapping[str, object],
-    *,
-    remaining_qty: Decimal,
-    original_qty: Decimal,
-) -> RuntimeLedgerFill | dict[str, object]:
-    ratio = remaining_qty / original_qty if original_qty > 0 else Decimal("1")
-    price = _ledger_row_decimal(
-        row,
-        "avg_fill_price",
-        "filled_avg_price",
-        "filled_average_price",
-        "average_fill_price",
-        "fill_price",
-        "filled_price",
-    )
-    filled_notional = remaining_qty * price if price is not None else None
-    cost_amount = _ledger_row_decimal(row, "cost_amount")
-    adjusted_cost = cost_amount * ratio if cost_amount is not None else None
-    if isinstance(row, RuntimeLedgerFill):
-        return replace(
-            row,
-            filled_qty=remaining_qty,
-            filled_notional=filled_notional,
-            cost_amount=adjusted_cost,
-        )
-    adjusted = dict(row)
-    adjusted["filled_qty"] = remaining_qty
-    if "filled_qty_delta" in adjusted:
-        adjusted["filled_qty_delta"] = remaining_qty
-    if filled_notional is not None:
-        adjusted["filled_notional"] = filled_notional
-        if "filled_notional_delta" in adjusted:
-            adjusted["filled_notional_delta"] = filled_notional
-    if adjusted_cost is not None:
-        adjusted["cost_amount"] = adjusted_cost
-    return adjusted
-
-
-def _active_carry_in_ledger_rows(
-    rows: Sequence[RuntimeLedgerFill | Mapping[str, object]],
-    *,
-    bucket_start: datetime,
-) -> tuple[list[RuntimeLedgerFill | dict[str, object]], set[str], set[str]]:
-    lots_by_key: dict[
-        tuple[str | None, str | None, str | None],
-        list[dict[str, object]],
-    ] = {}
-    for row in sorted(
-        rows,
-        key=lambda item: (
-            _ledger_row_time(item) or bucket_start,
-            _ledger_row_order_id(item) or "",
-        ),
-    ):
-        event_time = _ledger_row_time(row)
-        if event_time is None or event_time >= bucket_start:
-            continue
-        if _ledger_row_event_type(row) not in _RUNTIME_LEDGER_FILL_EVENTS:
-            continue
-        side_sign = _ledger_row_side_sign(row)
-        if side_sign == 0:
-            continue
-        qty = _ledger_row_decimal(
-            row, "filled_qty", "filled_quantity", "qty", "quantity"
-        )
-        price = _ledger_row_decimal(
-            row,
-            "avg_fill_price",
-            "filled_avg_price",
-            "filled_average_price",
-            "average_fill_price",
-            "fill_price",
-            "filled_price",
-        )
-        if qty is None or qty <= 0 or price is None or price <= 0:
-            continue
-        key = _ledger_row_position_key(row)
-        lots = lots_by_key.setdefault(key, [])
-        remaining_qty = qty
-        for lot in list(lots):
-            if remaining_qty <= 0:
-                break
-            if cast(Decimal, lot["side_sign"]) == side_sign:
-                continue
-            lot_qty = cast(Decimal, lot["remaining_qty"])
-            close_qty = min(remaining_qty, lot_qty)
-            lot["remaining_qty"] = lot_qty - close_qty
-            remaining_qty -= close_qty
-            if cast(Decimal, lot["remaining_qty"]) <= 0:
-                lots.remove(lot)
-        if remaining_qty > 0:
-            lots.append(
-                {
-                    "row": row,
-                    "side_sign": side_sign,
-                    "remaining_qty": remaining_qty,
-                    "original_qty": qty,
-                }
-            )
-
-    active_rows: list[RuntimeLedgerFill | dict[str, object]] = []
-    active_order_ids: set[str] = set()
-    active_decision_ids: set[str] = set()
-    for lots in lots_by_key.values():
-        for lot in lots:
-            remaining_qty = cast(Decimal, lot["remaining_qty"])
-            original_qty = cast(Decimal, lot["original_qty"])
-            if remaining_qty <= 0 or original_qty <= 0:
-                continue
-            row = cast(RuntimeLedgerFill | Mapping[str, object], lot["row"])
-            adjusted = _adjust_carry_in_lot_row(
-                row,
-                remaining_qty=remaining_qty,
-                original_qty=original_qty,
-            )
-            active_rows.append(adjusted)
-            if (order_id := _ledger_row_order_id(adjusted)) is not None:
-                active_order_ids.add(order_id)
-            if (decision_id := _ledger_row_decision_id(adjusted)) is not None:
-                active_decision_ids.add(decision_id)
-    linked_lifecycle_rows: list[RuntimeLedgerFill | dict[str, object]] = []
-    for row in rows:
-        event_time = _ledger_row_time(row)
-        if event_time is None or event_time >= bucket_start:
-            continue
-        if _ledger_row_event_type(row) in _RUNTIME_LEDGER_FILL_EVENTS:
-            continue
-        order_id = _ledger_row_order_id(row)
-        decision_id = _ledger_row_decision_id(row)
-        if (order_id is not None and order_id in active_order_ids) or (
-            decision_id is not None and decision_id in active_decision_ids
-        ):
-            linked_lifecycle_rows.append(dict(row) if isinstance(row, Mapping) else row)
-    return [*linked_lifecycle_rows, *active_rows], active_order_ids, active_decision_ids
-
-
-def _filter_carry_in_source_rows_for_active_lots(
-    rows: list[dict[str, object]] | None,
-    *,
-    active_order_ids: set[str],
-    active_decision_ids: set[str],
-) -> list[dict[str, object]] | None:
-    if not rows or (not active_order_ids and not active_decision_ids):
-        return None
-    filtered = [
-        row
-        for row in rows
-        if (
-            (order_id := _runtime_order_id(row)) is not None
-            and order_id in active_order_ids
-        )
-        or bool(_runtime_source_decision_ids([row]) & active_decision_ids)
-    ]
-    return filtered or None
-
-
-def _flat_start_position_snapshot_from_cursor(
-    cur: Any,
-    *,
-    account_label: str,
-    window_start: datetime,
-) -> dict[str, object] | None:
-    window_start_utc = (
-        window_start.astimezone(timezone.utc)
-        if window_start.tzinfo is not None
-        else window_start.replace(tzinfo=timezone.utc)
-    )
-    cur.execute(
-        """
-        with before_snapshot as (
-            select
-                id::text,
-                as_of,
-                positions,
-                'latest_before_window_start' as snapshot_source
-            from position_snapshots
-            where alpaca_account_label = %s
-              and as_of <= %s
-            order by as_of desc
-            limit 1
-        ),
-        after_snapshot as (
-            select
-                id::text,
-                as_of,
-                positions,
-                'first_after_window_start' as snapshot_source
-            from position_snapshots
-            where alpaca_account_label = %s
-              and as_of > %s
-              and as_of <= %s
-            order by as_of asc
-            limit 1
-        )
-        select id, as_of, positions, snapshot_source, 0 as snapshot_priority
-        from before_snapshot
-        union all
-        select id, as_of, positions, snapshot_source, 1 as snapshot_priority
-        from after_snapshot
-        order by snapshot_priority
-        limit 1
-        """,
-        (
-            account_label,
-            window_start_utc,
-            account_label,
-            window_start_utc,
-            window_start_utc
-            + timedelta(
-                seconds=RUNTIME_LEDGER_FLAT_START_SNAPSHOT_AFTER_START_GRACE_SECONDS
-            ),
-        ),
-    )
-    rows = list(cur.fetchall())
-    if not rows:
-        return None
-    snapshot_id, snapshot_as_of, positions, snapshot_source, _snapshot_priority = rows[
-        0
-    ]
-    parsed_as_of = _parse_dt_or_none(snapshot_as_of)
-    if parsed_as_of is None:
-        return None
-    offset_seconds = int((parsed_as_of - window_start_utc).total_seconds())
-    blockers: list[str] = []
-    if offset_seconds < -RUNTIME_LEDGER_FLAT_START_SNAPSHOT_STALE_SECONDS:
-        blockers.append("runtime_ledger_flat_start_position_snapshot_stale")
-    position_count = _position_snapshot_open_position_count(positions)
-    if position_count is None:
-        blockers.append("runtime_ledger_flat_start_position_snapshot_positions_invalid")
-        position_count = 0
-    if position_count > 0:
-        blockers.append("runtime_ledger_flat_start_position_snapshot_not_flat")
-    return {
-        "schema_version": "torghut.runtime-ledger-flat-start-position-snapshot.v1",
-        "scope": "account_position_snapshot_at_runtime_window_start",
-        "account_label": account_label,
-        "snapshot_id": str(snapshot_id),
-        "snapshot_as_of": parsed_as_of.isoformat(),
-        "snapshot_source": str(snapshot_source or "position_snapshots"),
-        "snapshot_offset_seconds": offset_seconds,
-        "flat": position_count == 0 and not blockers,
-        "position_count": position_count,
-        "blockers": blockers,
-    }
-
-
-def _runtime_order_rows_for_bucket(
-    *,
-    bucket: RuntimeLedgerBucket,
-    execution_rows: list[dict[str, object]],
-    order_lifecycle_rows: list[dict[str, object]] | None,
-) -> list[dict[str, object]]:
-    bucket_execution_order_ids = {
-        order_id
-        for row in execution_rows
-        if _runtime_source_row_in_bucket(row, bucket=bucket)
-        and _execution_row_has_fill(row)
-        and (order_id := _runtime_order_id(row)) is not None
-    }
-    bucket_fill_order_ids = {
-        order_id
-        for row in order_lifecycle_rows or []
-        if _runtime_source_row_in_bucket(row, bucket=bucket)
-        and _runtime_ledger_event_type({str(key): value for key, value in row.items()})
-        in _RUNTIME_LEDGER_FILL_EVENTS
-        and (order_id := _runtime_order_id(row)) is not None
-    }
-    bucket_order_ids = bucket_execution_order_ids | bucket_fill_order_ids
-    rows: list[dict[str, object]] = []
-    for row in order_lifecycle_rows or []:
-        if not _runtime_source_row_in_bucket(row, bucket=bucket):
-            continue
-        order_id = _runtime_order_id(row)
-        if bucket_order_ids and order_id not in bucket_order_ids:
-            continue
-        rows.append(dict(row))
-    return rows
-
-
-def _runtime_decision_rows_for_bucket(
-    *,
-    bucket: RuntimeLedgerBucket,
-    decision_lifecycle_rows: list[dict[str, object]] | None,
-    source_rows: Sequence[Mapping[str, object]],
-) -> list[dict[str, object]]:
-    decision_ids = _runtime_source_decision_ids(source_rows)
-    rows: list[dict[str, object]] = []
-    for row in decision_lifecycle_rows or []:
-        if not _runtime_source_row_in_bucket(row, bucket=bucket):
-            continue
-        row_ids = _runtime_source_decision_ids([row])
-        if decision_ids and not (row_ids & decision_ids):
-            continue
-        rows.append(dict(row))
-    return rows
-
-
-def _runtime_decision_rows_before_bucket(
-    *,
-    bucket: RuntimeLedgerBucket,
-    decision_lifecycle_rows: list[dict[str, object]] | None,
-    source_rows: Sequence[Mapping[str, object]],
-) -> list[dict[str, object]]:
-    decision_ids = _runtime_source_decision_ids(source_rows)
-    rows: list[dict[str, object]] = []
-    for row in decision_lifecycle_rows or []:
-        if not _runtime_source_row_before_bucket(row, bucket=bucket):
-            continue
-        row_ids = _runtime_source_decision_ids([row])
-        if decision_ids and not (row_ids & decision_ids):
-            continue
-        rows.append(dict(row))
-    return rows
-
-
-def _runtime_unlinked_order_rows_for_bucket(
-    *,
-    bucket: RuntimeLedgerBucket,
-    unlinked_order_lifecycle_rows: list[dict[str, object]] | None,
-    bucket_order_ids: set[str],
-) -> list[dict[str, object]]:
-    rows: list[dict[str, object]] = []
-    for row in unlinked_order_lifecycle_rows or []:
-        if _runtime_source_row_in_bucket(row, bucket=bucket):
-            rows.append(dict(row))
-            continue
-        row_symbol = _runtime_source_row_symbol(row)
-        if (
-            bucket.symbol is not None
-            and row_symbol is not None
-            and row_symbol != bucket.symbol
-        ):
-            continue
-        order_id = _runtime_order_id(row)
-        if order_id is not None and order_id in bucket_order_ids:
-            rows.append(dict(row))
-    return rows
-
-
-def _event_sourced_fill_economics_order_ids(
-    rows: Sequence[Mapping[str, object]],
-) -> set[str]:
-    order_ids: set[str] = set()
-    for row in rows:
-        normalized = {str(key): value for key, value in row.items()}
-        if _runtime_ledger_event_type(normalized) not in _RUNTIME_LEDGER_FILL_EVENTS:
-            continue
-        if (
-            _runtime_lifecycle_ledger_row(
-                normalized,
-                event_type=_runtime_ledger_event_type(normalized),
-                require_complete_fill=True,
-            )
-            is None
-        ):
-            continue
-        order_id = _runtime_order_id(normalized)
-        if order_id is not None:
-            order_ids.add(order_id)
-    return order_ids
-
-
-def _source_backed_fill_lifecycle_rows(
-    rows: Sequence[Mapping[str, object]],
-) -> list[dict[str, object]]:
-    source_backed_rows: list[dict[str, object]] = []
-    for row in rows:
-        normalized = _with_canonical_runtime_source_refs(row)
-        if _runtime_ledger_event_type(normalized) not in _RUNTIME_LEDGER_FILL_EVENTS:
-            continue
-        if _runtime_order_id(normalized) is None:
-            continue
-        if not _source_identifier_values(
-            [normalized],
-            *_EXECUTION_ORDER_EVENT_REF_KEYS,
-        ):
-            continue
-        if not _source_identifier_values([normalized], *_SOURCE_WINDOW_REF_KEYS):
-            continue
-        if not _source_offset_values([normalized]):
-            continue
-        source_backed_rows.append(dict(normalized))
-    return source_backed_rows
-
-
-def _source_backed_order_lifecycle_rows(
-    rows: Sequence[Mapping[str, object]],
-) -> list[dict[str, object]]:
-    source_backed_rows: list[dict[str, object]] = []
-    for row in rows:
-        normalized = _with_canonical_runtime_source_refs(row)
-        if (
-            _runtime_ledger_event_type(normalized)
-            not in _RUNTIME_LEDGER_ORDER_LIFECYCLE_EVENTS
-        ):
-            continue
-        if _runtime_order_id(normalized) is None:
-            continue
-        if not _source_identifier_values(
-            [normalized],
-            *_EXECUTION_ORDER_EVENT_REF_KEYS,
-        ):
-            continue
-        if not _source_identifier_values([normalized], *_SOURCE_WINDOW_REF_KEYS):
-            continue
-        if not _source_offset_values([normalized]):
-            continue
-        source_backed_rows.append(dict(normalized))
-    return source_backed_rows
-
-
-def _required_order_lifecycle_source_row_count(
-    rows: Sequence[Mapping[str, object]],
-    *,
-    expected_order_ids: set[str],
-) -> int:
-    if not expected_order_ids:
-        return 0
-    count = 0
-    for row in rows:
-        normalized = {str(key): value for key, value in row.items()}
-        if (
-            _runtime_ledger_event_type(normalized)
-            not in _RUNTIME_LEDGER_ORDER_LIFECYCLE_EVENTS
-        ):
-            continue
-        order_id = _runtime_order_id(normalized)
-        if order_id is None or order_id not in expected_order_ids:
-            continue
-        count += 1
-    return max(count, len(expected_order_ids))
-
-
-def _source_backed_fill_lifecycle_order_ids(
-    rows: Sequence[Mapping[str, object]],
-) -> set[str]:
-    return {
-        order_id
-        for row in _source_backed_fill_lifecycle_rows(rows)
-        if (order_id := _runtime_order_id(row)) is not None
-    }
-
-
-def _source_backed_submitted_lifecycle_order_ids(
-    rows: Sequence[Mapping[str, object]],
-) -> set[str]:
-    order_ids: set[str] = set()
-    for row in _source_backed_order_lifecycle_rows(rows):
-        normalized = {str(key): value for key, value in row.items()}
-        if _runtime_ledger_event_type(normalized) != "order_submitted":
-            continue
-        order_id = _runtime_order_id(normalized)
-        if order_id is not None:
-            order_ids.add(order_id)
-    return order_ids
-
-
-def _execution_fill_economics_order_ids(
-    rows: Sequence[Mapping[str, object]],
-) -> set[str]:
-    order_ids: set[str] = set()
-    for row in rows:
-        if not _execution_row_has_fill(row):
-            continue
-        order_id = _runtime_order_id(row)
-        if order_id is None:
-            continue
-        filled_notional = _first_positive_decimal(
-            row, "filled_notional", "notional", "turnover_notional"
-        ) or Decimal("0")
-        if filled_notional <= 0:
-            price = _first_positive_decimal(
-                row,
-                "avg_fill_price",
-                "filled_avg_price",
-                "filled_average_price",
-                "average_fill_price",
-                "fill_price",
-                "filled_price",
-            )
-            qty = _first_positive_decimal(
-                row, "filled_qty", "filled_quantity", "qty", "quantity"
-            )
-            if price is None or qty is None or price <= 0 or qty <= 0:
-                continue
-            filled_notional = price * qty
-        cost_amount = _runtime_execution_cost_amount(
-            row,
-            filled_notional=filled_notional,
-            side=_first_text(row, "side", "order_side") or "",
-            filled_qty=(
-                _first_positive_decimal(
-                    row, "filled_qty", "filled_quantity", "qty", "quantity"
-                )
-                or Decimal("0")
-            ),
-        )
-        cost_basis = _runtime_execution_cost_basis(
-            row,
-            cost_amount=cost_amount,
-            side=_first_text(row, "side", "order_side") or "",
-            filled_qty=(
-                _first_positive_decimal(
-                    row, "filled_qty", "filled_quantity", "qty", "quantity"
-                )
-                or Decimal("0")
-            ),
-            filled_notional=filled_notional,
-        )
-        if cost_amount is not None and cost_amount >= 0 and cost_basis is not None:
-            order_ids.add(order_id)
-    return order_ids
-
-
-def _execution_tca_order_ids(
-    rows: Sequence[Mapping[str, object]],
-) -> set[str]:
-    order_ids: set[str] = set()
-    for row in rows:
-        if not _execution_row_has_fill(row):
-            continue
-        if (
-            _first_text(
-                row,
-                "execution_tca_metric_id",
-                "execution_tca_metric_ref",
-                "execution_tca_id",
-                "tca_metric_id",
-                "tca_id",
-            )
-            is None
-        ):
-            continue
-        order_id = _runtime_order_id(row)
-        if order_id is not None:
-            order_ids.add(order_id)
-    return order_ids
-
-
-def _runtime_source_context_for_bucket(
-    *,
-    bucket: RuntimeLedgerBucket,
-    execution_rows: list[dict[str, object]],
-    decision_lifecycle_rows: list[dict[str, object]] | None,
-    order_lifecycle_rows: list[dict[str, object]] | None,
-    unlinked_order_lifecycle_rows: list[dict[str, object]] | None,
-    carry_in_execution_rows: list[dict[str, object]] | None = None,
-    carry_in_decision_lifecycle_rows: list[dict[str, object]] | None = None,
-    carry_in_order_lifecycle_rows: list[dict[str, object]] | None = None,
-) -> dict[str, object]:
-    target_order_rows = _runtime_order_rows_for_bucket(
-        bucket=bucket,
-        execution_rows=execution_rows,
-        order_lifecycle_rows=order_lifecycle_rows,
-    )
-    bucket_carry_in_order_rows = [
-        dict(row)
-        for row in carry_in_order_lifecycle_rows or []
-        if _runtime_source_row_before_bucket(row, bucket=bucket)
-    ]
-    bucket_order_rows = [*bucket_carry_in_order_rows, *target_order_rows]
-    bucket_order_ids = {
-        order_id
-        for row in bucket_order_rows
-        if (order_id := _runtime_order_id(row)) is not None
-    }
-    target_execution_rows = [
-        dict(row)
-        for row in execution_rows
-        if _execution_row_has_fill(row)
-        and (
-            _runtime_source_row_in_bucket(row, bucket=bucket)
-            or (
-                (order_id := _runtime_order_id(row)) is not None
-                and order_id in bucket_order_ids
-            )
-        )
-    ]
-    bucket_carry_in_execution_rows = [
-        dict(row)
-        for row in carry_in_execution_rows or []
-        if _execution_row_has_fill(row)
-        and _runtime_source_row_before_bucket(row, bucket=bucket)
-    ]
-    bucket_execution_rows = [*bucket_carry_in_execution_rows, *target_execution_rows]
-    target_decision_rows = _runtime_decision_rows_for_bucket(
-        bucket=bucket,
-        decision_lifecycle_rows=decision_lifecycle_rows,
-        source_rows=[*bucket_execution_rows, *bucket_order_rows],
-    )
-    carry_in_decision_rows = _runtime_decision_rows_before_bucket(
-        bucket=bucket,
-        decision_lifecycle_rows=carry_in_decision_lifecycle_rows,
-        source_rows=[*bucket_execution_rows, *bucket_order_rows],
-    )
-    bucket_decision_rows = [*carry_in_decision_rows, *target_decision_rows]
-    bucket_unlinked_order_rows = _runtime_unlinked_order_rows_for_bucket(
-        bucket=bucket,
-        unlinked_order_lifecycle_rows=unlinked_order_lifecycle_rows,
-        bucket_order_ids=bucket_order_ids,
-    )
-    source_rows = [*bucket_execution_rows, *bucket_decision_rows, *bucket_order_rows]
-    source_backed_fill_lifecycle_rows = _source_backed_fill_lifecycle_rows(
-        bucket_order_rows
-    )
-    source_backed_order_lifecycle_rows = _source_backed_order_lifecycle_rows(
-        bucket_order_rows
-    )
-    expected_execution_fill_order_ids = set()
-    for row in bucket_execution_rows:
-        if (
-            _execution_row_has_fill(row)
-            and (order_id := _runtime_order_id(row)) is not None
-        ):
-            expected_execution_fill_order_ids.add(order_id)
-    for row in bucket_order_rows:
-        if (order_id := _runtime_order_id(row)) is not None:
-            expected_execution_fill_order_ids.add(order_id)
-    source_backed_fill_lifecycle_order_ids = _source_backed_fill_lifecycle_order_ids(
-        source_backed_fill_lifecycle_rows
-    )
-    source_backed_submitted_lifecycle_order_ids = (
-        _source_backed_submitted_lifecycle_order_ids(source_backed_order_lifecycle_rows)
-    )
-    event_sourced_fill_order_ids = _event_sourced_fill_economics_order_ids(
-        source_backed_fill_lifecycle_rows
-    )
-    execution_fill_order_ids = _execution_fill_economics_order_ids(
-        bucket_execution_rows
-    )
-    execution_tca_order_ids = _execution_tca_order_ids(bucket_execution_rows)
-    execution_tca_required = any(
-        _execution_row_has_fill(row)
-        and any(
-            key in row
-            for key in (
-                "execution_tca_metric_id",
-                "execution_tca_metric_ref",
-                "execution_tca_id",
-                "tca_metric_id",
-                "tca_id",
-            )
-        )
-        for row in bucket_execution_rows
-    )
-    order_feed_fill_economics_complete = bool(
-        expected_execution_fill_order_ids
-    ) and expected_execution_fill_order_ids.issubset(event_sourced_fill_order_ids)
-    execution_fill_economics_complete = bool(
-        expected_execution_fill_order_ids
-    ) and expected_execution_fill_order_ids.issubset(execution_fill_order_ids)
-    execution_tca_complete = bool(
-        expected_execution_fill_order_ids
-    ) and expected_execution_fill_order_ids.issubset(execution_tca_order_ids)
-    execution_tca_satisfied = (not execution_tca_required) or execution_tca_complete
-    source_backed_fill_lifecycle_complete = bool(
-        expected_execution_fill_order_ids
-    ) and expected_execution_fill_order_ids.issubset(
-        source_backed_fill_lifecycle_order_ids
-    )
-    source_backed_submitted_lifecycle_complete = bool(
-        expected_execution_fill_order_ids
-    ) and expected_execution_fill_order_ids.issubset(
-        source_backed_submitted_lifecycle_order_ids
-    )
-    source_backed_order_lifecycle_complete = (
-        source_backed_fill_lifecycle_complete
-        and source_backed_submitted_lifecycle_complete
-    )
-    source_authority_lifecycle_rows = (
-        source_backed_order_lifecycle_rows
-        if source_backed_order_lifecycle_complete
-        else source_backed_fill_lifecycle_rows
-    )
-    source_authority_lifecycle_times = [
-        event_time
-        for row in source_authority_lifecycle_rows
-        if (
-            event_time := _runtime_ledger_row_time(
-                {str(key): value for key, value in row.items()}
-            )
-        )
-        is not None
-    ]
-    source_window_ids = _source_identifier_values(
-        source_authority_lifecycle_rows,
-        *_SOURCE_WINDOW_REF_KEYS,
-    )
-    required_order_lifecycle_source_row_count = (
-        _required_order_lifecycle_source_row_count(
-            bucket_order_rows,
-            expected_order_ids=expected_execution_fill_order_ids,
-        )
-    )
-    execution_tca_metric_ids = _source_identifier_values(
-        bucket_execution_rows,
-        "execution_tca_metric_id",
-        "execution_tca_metric_ref",
-        "execution_tca_id",
-        "tca_metric_id",
-        "tca_id",
-    )
-    source_row_counts = {
-        "trade_decisions": len(bucket_decision_rows),
-        "executions": len(bucket_execution_rows),
-        "execution_order_events": required_order_lifecycle_source_row_count,
-    }
-    if execution_tca_metric_ids:
-        source_row_counts["execution_tca_metrics"] = len(execution_tca_metric_ids)
-    if source_window_ids:
-        source_row_counts["order_feed_source_windows"] = len(source_window_ids)
-    elif required_order_lifecycle_source_row_count > 0:
-        source_row_counts["order_feed_source_windows"] = 1
-    source_refs = [
-        ref
-        for table, ref in (
-            ("trade_decisions", "postgres:trade_decisions"),
-            ("executions", "postgres:executions"),
-            ("execution_tca_metrics", "postgres:execution_tca_metrics"),
-            ("execution_order_events", "postgres:execution_order_events"),
-            ("order_feed_source_windows", "postgres:order_feed_source_windows"),
-        )
-        if source_row_counts.get(table, 0) > 0
-    ]
-    source_materialization = None
-    authority_class = None
-    authority_reason = None
-    source_offsets = _source_offset_values(source_authority_lifecycle_rows)
-    source_window_status_counts = _source_window_status_counts(
-        source_authority_lifecycle_rows
-    )
-    source_window_classification_counts = _source_window_classification_counts(
-        source_authority_lifecycle_rows
-    )
-    source_window_gap_count = _source_window_gap_count(source_authority_lifecycle_rows)
-    source_window_gap_ranges = _source_window_gap_ranges(
-        source_authority_lifecycle_rows
-    )
-    execution_order_event_ids = _source_identifier_values(
-        source_authority_lifecycle_rows,
-        *_EXECUTION_ORDER_EVENT_REF_KEYS,
-    )
-    if source_offsets and execution_order_event_ids:
-        if order_feed_fill_economics_complete and execution_tca_satisfied:
-            source_materialization = "execution_order_events"
-            authority_class = "runtime_order_feed_execution_source"
-            authority_reason = "event_sourced_runtime_ledger_profit_proof"
-        elif (
-            execution_fill_economics_complete
-            and execution_tca_satisfied
-            and source_backed_order_lifecycle_complete
-        ):
-            source_materialization = "source_execution_lifecycle"
-            authority_class = "source_execution_lifecycle_materialized_runtime_ledger"
-            authority_reason = "source_execution_lifecycle_materialized_runtime_ledger"
-    return {
-        "execution_rows": bucket_execution_rows,
-        "decision_rows": bucket_decision_rows,
-        "order_rows": bucket_order_rows,
-        "unlinked_order_rows": bucket_unlinked_order_rows,
-        "source_rows": source_rows,
-        "source_window_ids": source_window_ids,
-        "source_window_start": (
-            min(source_authority_lifecycle_times)
-            if source_authority_lifecycle_times
-            else bucket.bucket_started_at
-        ),
-        "source_window_end": (
-            max(source_authority_lifecycle_times) + timedelta(microseconds=1)
-            if source_authority_lifecycle_times
-            else bucket.bucket_ended_at
-        ),
-        "source_row_counts": source_row_counts,
-        "source_refs": source_refs,
-        "trade_decision_ids": _source_identifier_values(
-            source_rows,
-            "trade_decision_id",
-            "decision_id",
-            "decision_hash",
-        ),
-        "execution_ids": _source_identifier_values(
-            [*bucket_execution_rows, *bucket_order_rows],
-            "execution_id",
-        ),
-        "execution_tca_metric_ids": execution_tca_metric_ids,
-        "execution_order_event_ids": execution_order_event_ids,
-        "source_window_status_counts": source_window_status_counts,
-        "source_window_classification_counts": source_window_classification_counts,
-        "source_window_gap_count": source_window_gap_count,
-        "source_window_gap_ranges": source_window_gap_ranges,
-        "source_offsets": source_offsets,
-        "source_materialization": source_materialization,
-        "authority_class": authority_class,
-        "authority_reason": authority_reason,
-        "order_feed_lifecycle_complete": source_backed_order_lifecycle_complete,
-        "fill_economics_complete": (
-            (order_feed_fill_economics_complete or execution_fill_economics_complete)
-            and execution_tca_satisfied
-        ),
-        "execution_tca_required": execution_tca_required,
-        "execution_tca_complete": execution_tca_complete,
-        "order_feed_fill_lifecycle_blockers": _order_feed_fill_lifecycle_blockers(
-            execution_rows=bucket_execution_rows,
-            order_lifecycle_rows=bucket_order_rows,
-            unlinked_order_lifecycle_rows=bucket_unlinked_order_rows,
-        ),
-        "source_account_labels": sorted(
-            {
-                str(row.get("source_account_label") or "").strip()
-                for row in source_rows
-                if str(row.get("source_account_label") or "").strip()
-            }
-        ),
-        "source_decision_mode_counts": _source_decision_mode_counts(source_rows),
-        "source_decision_profit_proof_eligible": (
-            _source_decision_rows_profit_proof_eligible(source_rows)
-        ),
-        "paper_route_target_notional_sizing": (
-            _source_decision_target_notional_sizing_summary(source_rows)
-        ),
-        "equity_denominator": _runtime_ledger_equity_denominator_from_rows(source_rows),
-    }
-
-
-def _source_activity_diagnostics_blockers(
-    diagnostics: Mapping[str, object],
-) -> list[str]:
-    blockers: list[str] = []
-
-    def add(blocker: str) -> None:
-        if blocker not in blockers:
-            blockers.append(blocker)
-
-    if _text_or_none(diagnostics.get("runtime_ledger_source_bucket_unavailable")):
-        add("runtime_ledger_source_bucket_unavailable")
-
-    decision_query_count = _nonnegative_int(
-        diagnostics.get("decision_rows_before_lineage_filter")
-    )
-    execution_query_count = _nonnegative_int(
-        diagnostics.get("execution_rows_before_lineage_filter")
-    )
-    order_query_count = _nonnegative_int(
-        diagnostics.get("order_lifecycle_rows_before_lineage_filter")
-    )
-    decision_lineage_count = _nonnegative_int(
-        diagnostics.get("decision_rows_after_lineage_filter")
-    )
-    execution_lineage_count = _nonnegative_int(
-        diagnostics.get("execution_rows_after_lineage_filter")
-    )
-    execution_tca_lineage_count = _nonnegative_int(
-        diagnostics.get("execution_tca_rows_after_lineage_filter")
-    )
-    source_bucket_count = _nonnegative_int(
-        diagnostics.get("runtime_ledger_source_bucket_count")
-    )
-    source_bucket_profit_proof_count = _nonnegative_int(
-        diagnostics.get("runtime_ledger_source_bucket_profit_proof_count")
-    )
-    unlinked_fill_lifecycle_count = _nonnegative_int(
-        diagnostics.get("order_feed_unlinked_fill_lifecycle_count")
-    )
-    if "order_feed_unlinked_strategy_fill_lifecycle_count" in diagnostics:
-        strategy_unlinked_fill_lifecycle_count = _nonnegative_int(
-            diagnostics.get("order_feed_unlinked_strategy_fill_lifecycle_count")
-        )
-    else:
-        strategy_unlinked_fill_lifecycle_count = unlinked_fill_lifecycle_count
-
-    if (decision_query_count or execution_query_count or order_query_count) and not (
-        decision_lineage_count or execution_lineage_count
-    ):
-        add("source_lineage_filter_excluded_activity")
-    elif not (
-        decision_query_count
-        or execution_query_count
-        or order_query_count
-        or source_bucket_count
-    ):
-        add("strategy_account_symbol_window_source_activity_missing")
-
-    if decision_lineage_count > 0 and execution_lineage_count <= 0:
-        add("execution_rows_missing_for_matched_decisions")
-    if execution_lineage_count > 0 and order_query_count <= 0:
-        add(ORDER_FEED_FILL_LIFECYCLE_MISSING_BLOCKER)
-    if execution_lineage_count > 0 and execution_tca_lineage_count <= 0:
-        add("execution_tca_rows_missing")
-    if strategy_unlinked_fill_lifecycle_count > 0:
-        add(ORDER_FEED_UNLINKED_FILL_LIFECYCLE_PRESENT_BLOCKER)
-    if source_bucket_count <= 0:
-        add("runtime_ledger_source_bucket_missing")
-    elif source_bucket_profit_proof_count <= 0:
-        add("runtime_ledger_source_bucket_profit_proof_missing")
-
-    for blocker in _metadata_text_list(
-        diagnostics.get("runtime_ledger_source_bucket_profit_proof_blockers")
-    ):
-        add(blocker)
-
-    for blocker in _metadata_text_list(
-        diagnostics.get("order_feed_fill_lifecycle_blockers")
-    ):
-        add(blocker)
-
-    return blockers
-
-
-def _runtime_execution_ledger_fill_from_row(
-    row: Mapping[str, object],
-    *,
-    order_lifecycle_rows: Sequence[Mapping[str, object]] | None,
-    event_sourced_fill_economics_order_ids: set[str] | None = None,
-) -> tuple[RuntimeLedgerFill | None, list[datetime]]:
-    computed_at = row.get("computed_at")
-    execution_event_at = row.get("execution_event_at")
-    execution_created_at = row.get("execution_created_at")
-    fill_event_at = (
-        execution_created_at
-        if isinstance(execution_created_at, datetime)
-        else computed_at
-    )
-    if isinstance(execution_event_at, datetime):
-        fill_event_at = execution_event_at
-    ledger_computed_at = (
-        fill_event_at if isinstance(fill_event_at, datetime) else computed_at
-    )
-    price = _first_positive_decimal(
-        row,
-        "avg_fill_price",
-        "filled_avg_price",
-        "filled_average_price",
-        "average_fill_price",
-        "fill_price",
-        "filled_price",
-    )
-    side = _first_text(row, "side", "order_side") or ""
-    signed_qty = _execution_signed_qty(
-        side=side,
-        qty=_first_positive_decimal(
-            row,
-            "filled_qty",
-            "filled_quantity",
-            "qty",
-            "quantity",
-        ),
-    )
-    symbol = str(row.get("symbol") or "").strip().upper()
-    if not symbol or not isinstance(ledger_computed_at, datetime):
-        return None, []
-    decision_id = _first_text(row, "decision_id", "trade_decision_id", "decision_hash")
-    order_id = _first_text(
-        row,
-        "order_id",
-        "alpaca_order_id",
-        "client_order_id",
-        "execution_correlation_id",
-    )
-    execution_policy_hash = _first_text(
-        row,
-        "execution_policy_hash",
-        "execution_policy_sha256",
-        "policy_hash",
-    ) or _first_payload_digest(
-        row,
-        "execution_policy",
-        "execution_policy_context",
-        "execution_advisor",
-        "_execution_advice_provenance",
-    )
-    cost_model_hash = _first_text(
-        row, "cost_model_hash", "fee_model_hash", "cost_model_sha256"
-    ) or _first_payload_digest(
-        row,
-        "cost_model",
-        "cost_model_config",
-        "transaction_cost_model",
-        "fee_model",
-        "fees_model",
-        "model",
-    )
-    lineage_hash = _first_text(
-        row,
-        "lineage_hash",
-        "candidate_lineage_hash",
-        "replay_lineage_hash",
-        "candidate_evaluation_key",
-    ) or _first_lineage_digest(row)
-    replay_data_hash = _first_text(
-        row,
-        "replay_data_hash",
-        "replay_tape_content_sha256",
-        "dataset_snapshot_hash",
-        "source_query_digest",
-    )
-    if price is None or price <= 0 or signed_qty == 0:
-        return None, []
-    event_sourced_order_ids = (
-        event_sourced_fill_economics_order_ids
-        if event_sourced_fill_economics_order_ids is not None
-        else _event_sourced_fill_economics_order_ids(order_lifecycle_rows or [])
-    )
-    if order_id is not None and order_id in event_sourced_order_ids:
-        return None, []
-    event_times = [ledger_computed_at]
-    if isinstance(fill_event_at, datetime):
-        event_times.append(fill_event_at)
-    filled_qty = abs(signed_qty)
-    filled_notional = filled_qty * price
-    cost_amount = _runtime_execution_cost_amount(
-        row,
-        filled_notional=filled_notional,
-        side=side,
-        filled_qty=filled_qty,
-    )
-    cost_basis = _runtime_execution_cost_basis(
-        row,
-        cost_amount=cost_amount,
-        side=side,
-        filled_qty=filled_qty,
-        filled_notional=filled_notional,
-    )
-    if _cost_basis_is_alpaca_fee_schedule(cost_basis):
-        cost_model_hash = _alpaca_2026_equity_fee_schedule_hash()
-    return (
-        RuntimeLedgerFill(
-            executed_at=fill_event_at
-            if isinstance(fill_event_at, datetime)
-            else ledger_computed_at,
-            event_type="fill",
-            decision_id=decision_id,
-            order_id=order_id,
-            execution_policy_hash=execution_policy_hash,
-            cost_model_hash=cost_model_hash,
-            lineage_hash=lineage_hash,
-            replay_data_hash=replay_data_hash,
-            side=side,
-            filled_qty=filled_qty,
-            avg_fill_price=price,
-            filled_notional=filled_notional,
-            cost_amount=cost_amount,
-            cost_basis=cost_basis,
-            account_label=str(row.get("account_label") or "") or None,
-            strategy_id=str(row.get("strategy_id") or "") or None,
-            symbol=symbol,
-        ),
-        event_times,
-    )
-
-
-def _build_realized_strategy_pnl_rows(
-    execution_rows: list[dict[str, object]],
-    *,
-    decision_lifecycle_rows: list[dict[str, object]] | None = None,
-    order_lifecycle_rows: list[dict[str, object]] | None = None,
-    unlinked_order_lifecycle_rows: list[dict[str, object]] | None = None,
-    carry_in_execution_rows: list[dict[str, object]] | None = None,
-    carry_in_decision_lifecycle_rows: list[dict[str, object]] | None = None,
-    carry_in_order_lifecycle_rows: list[dict[str, object]] | None = None,
-    flat_start_position_snapshot: Mapping[str, object] | None = None,
-    allow_authoritative_runtime_ledger_materialization: bool = False,
-    split_mixed_source_decision_modes: bool = True,
-) -> list[dict[str, object]]:
-    flat_start_authority = _flat_start_position_snapshot_authority(
-        flat_start_position_snapshot
-    )
-    if flat_start_authority is not None:
-        carry_in_execution_rows = None
-        carry_in_decision_lifecycle_rows = None
-        carry_in_order_lifecycle_rows = None
-    if split_mixed_source_decision_modes:
-        partitions = _partition_runtime_source_rows_by_decision_mode(
-            execution_rows=execution_rows,
-            decision_lifecycle_rows=decision_lifecycle_rows,
-            order_lifecycle_rows=order_lifecycle_rows,
-            unlinked_order_lifecycle_rows=unlinked_order_lifecycle_rows,
-            carry_in_execution_rows=carry_in_execution_rows,
-            carry_in_decision_lifecycle_rows=carry_in_decision_lifecycle_rows,
-            carry_in_order_lifecycle_rows=carry_in_order_lifecycle_rows,
-        )
-        if partitions is not None:
-            partition_rows: list[dict[str, object]] = []
-            for (
-                mode_key,
-                partition_execution_rows,
-                partition_decision_rows,
-                partition_order_rows,
-                partition_unlinked_order_rows,
-                partition_carry_in_execution_rows,
-                partition_carry_in_decision_rows,
-                partition_carry_in_order_rows,
-            ) in partitions:
-                for row in _build_realized_strategy_pnl_rows(
-                    partition_execution_rows,
-                    decision_lifecycle_rows=partition_decision_rows,
-                    order_lifecycle_rows=partition_order_rows,
-                    unlinked_order_lifecycle_rows=partition_unlinked_order_rows,
-                    carry_in_execution_rows=partition_carry_in_execution_rows,
-                    carry_in_decision_lifecycle_rows=partition_carry_in_decision_rows,
-                    carry_in_order_lifecycle_rows=partition_carry_in_order_rows,
-                    flat_start_position_snapshot=flat_start_position_snapshot,
-                    allow_authoritative_runtime_ledger_materialization=(
-                        allow_authoritative_runtime_ledger_materialization
-                    ),
-                    split_mixed_source_decision_modes=False,
-                ):
-                    row["source_decision_mode_partition"] = mode_key
-                    bucket = row.get("runtime_ledger_bucket")
-                    if isinstance(bucket, Mapping):
-                        row["runtime_ledger_bucket"] = {
-                            **dict(bucket),
-                            "source_decision_mode_partition": mode_key,
-                        }
-                    partition_rows.append(row)
-            return partition_rows
-
-    if not (
-        execution_rows
-        or decision_lifecycle_rows
-        or carry_in_execution_rows
-        or carry_in_decision_lifecycle_rows
-    ):
-        return []
-
-    ledger_rows: list[RuntimeLedgerFill | dict[str, object]] = []
-    carry_in_ledger_rows: list[RuntimeLedgerFill | dict[str, object]] = []
-    event_times: list[datetime] = []
-    execution_order_ids = {
-        order_id for row in execution_rows if (order_id := _runtime_order_id(row))
-    }
-    execution_ids = {
-        execution_id
-        for row in execution_rows
-        if (execution_id := _first_text(row, "execution_id"))
-    }
-    execution_symbols = {
-        symbol
-        for row in execution_rows
-        if (symbol := _runtime_source_row_symbol(row)) is not None
-    }
-    decision_lifecycle_ids = _runtime_source_decision_ids(decision_lifecycle_rows or [])
-    carry_in_execution_order_ids = {
-        order_id
-        for row in carry_in_execution_rows or []
-        if (order_id := _runtime_order_id(row))
-    }
-    carry_in_execution_ids = {
-        execution_id
-        for row in carry_in_execution_rows or []
-        if (execution_id := _first_text(row, "execution_id"))
-    }
-    carry_in_execution_symbols = {
-        symbol
-        for row in carry_in_execution_rows or []
-        if (symbol := _runtime_source_row_symbol(row)) is not None
-    }
-    carry_in_decision_lifecycle_ids = _runtime_source_decision_ids(
-        carry_in_decision_lifecycle_rows or []
-    )
-    execution_id_by_order_id = _execution_id_by_order_id(execution_rows)
-    carry_in_execution_id_by_order_id = _execution_id_by_order_id(
-        carry_in_execution_rows or []
-    )
-    execution_side_by_order_id = _execution_side_by_order_id(execution_rows)
-    carry_in_execution_side_by_order_id = _execution_side_by_order_id(
-        carry_in_execution_rows or []
-    )
-    order_lifecycle_rows = [
-        _with_canonical_runtime_source_refs(
-            _with_linked_execution_id(
-                row,
-                execution_id_by_order_id=execution_id_by_order_id,
-                execution_side_by_order_id=execution_side_by_order_id,
-            )
-        )
-        for row in order_lifecycle_rows or []
-    ]
-    carry_in_order_lifecycle_rows = [
-        _with_canonical_runtime_source_refs(
-            _with_linked_execution_id(
-                row,
-                execution_id_by_order_id=carry_in_execution_id_by_order_id,
-                execution_side_by_order_id=carry_in_execution_side_by_order_id,
-            )
-        )
-        for row in carry_in_order_lifecycle_rows or []
-    ]
-    event_sourced_fill_economics_order_ids = _event_sourced_fill_economics_order_ids(
-        order_lifecycle_rows
-    )
-    for row in decision_lifecycle_rows or []:
-        lifecycle_row = _runtime_lifecycle_ledger_row(row, event_type="decision")
-        if lifecycle_row is None:
-            continue
-        ledger_rows.append(lifecycle_row)
-        event_time = lifecycle_row.get("executed_at")
-        if isinstance(event_time, datetime):
-            event_times.append(event_time)
-    for row in order_lifecycle_rows:
-        row_order_id = _runtime_order_id(row)
-        row_execution_id = _first_text(row, "execution_id")
-        row_symbol = _runtime_source_row_symbol(row)
-        if (
-            (execution_order_ids or execution_ids or execution_symbols)
-            and row_order_id not in execution_order_ids
-            and row_execution_id not in execution_ids
-            and row_symbol not in execution_symbols
-        ):
-            continue
-        if (
-            not (execution_order_ids or execution_ids or execution_symbols)
-            and decision_lifecycle_ids
-        ):
-            row_decision_ids = _runtime_source_decision_ids([row])
-            if row_decision_ids and not (row_decision_ids & decision_lifecycle_ids):
-                continue
-        event_type = _runtime_ledger_event_type(
-            {str(key): value for key, value in row.items()}
-        )
-        if event_type in _RUNTIME_LEDGER_FILL_EVENTS:
-            lifecycle_row = _runtime_lifecycle_ledger_row(
-                row,
-                event_type=event_type,
-                require_complete_fill=True,
-            )
-            if lifecycle_row is None:
-                if (
-                    row_order_id in execution_order_ids
-                    or row_execution_id in execution_ids
-                    or (
-                        not execution_order_ids
-                        and not execution_ids
-                        and not execution_symbols
-                        and _source_authority_order_event_row(row)
-                    )
-                ):
-                    lifecycle_row = _runtime_lifecycle_ledger_row(
-                        row,
-                        event_type=event_type,
-                    )
-                    if lifecycle_row is not None:
-                        lifecycle_row["source"] = "order_feed_lifecycle"
-            if lifecycle_row is not None:
-                ledger_rows.append(lifecycle_row)
-                event_time = lifecycle_row.get("executed_at")
-                if isinstance(event_time, datetime):
-                    event_times.append(event_time)
-                continue
-            event_time = _runtime_ledger_row_time(row)
-            if isinstance(event_time, datetime):
-                event_times.append(event_time)
-            continue
-        lifecycle_row = _runtime_lifecycle_ledger_row(
-            row,
-            event_type=event_type,
-        )
-        if lifecycle_row is None:
-            continue
-        ledger_rows.append(lifecycle_row)
-        event_time = lifecycle_row.get("executed_at")
-        if isinstance(event_time, datetime):
-            event_times.append(event_time)
-    for row in execution_rows:
-        ledger_fill, ledger_event_times = _runtime_execution_ledger_fill_from_row(
-            row,
-            order_lifecycle_rows=order_lifecycle_rows,
-            event_sourced_fill_economics_order_ids=(
-                event_sourced_fill_economics_order_ids
-            ),
-        )
-        if ledger_fill is None:
-            continue
-        ledger_rows.append(ledger_fill)
-        event_times.extend(ledger_event_times)
-
-    for row in carry_in_decision_lifecycle_rows or []:
-        lifecycle_row = _runtime_lifecycle_ledger_row(row, event_type="decision")
-        if lifecycle_row is not None:
-            carry_in_ledger_rows.append(lifecycle_row)
-    for row in carry_in_order_lifecycle_rows:
-        row_order_id = _runtime_order_id(row)
-        row_execution_id = _first_text(row, "execution_id")
-        row_symbol = _runtime_source_row_symbol(row)
-        if (
-            (
-                carry_in_execution_order_ids
-                or carry_in_execution_ids
-                or carry_in_execution_symbols
-            )
-            and row_order_id not in carry_in_execution_order_ids
-            and row_execution_id not in carry_in_execution_ids
-            and row_symbol not in carry_in_execution_symbols
-        ):
-            continue
-        if (
-            not (
-                carry_in_execution_order_ids
-                or carry_in_execution_ids
-                or carry_in_execution_symbols
-            )
-            and carry_in_decision_lifecycle_ids
-        ):
-            row_decision_ids = _runtime_source_decision_ids([row])
-            if row_decision_ids and not (
-                row_decision_ids & carry_in_decision_lifecycle_ids
-            ):
-                continue
-        event_type = _runtime_ledger_event_type(
-            {str(key): value for key, value in row.items()}
-        )
-        lifecycle_row = _runtime_lifecycle_ledger_row(
-            row,
-            event_type=event_type,
-            require_complete_fill=event_type in _RUNTIME_LEDGER_FILL_EVENTS,
-        )
-        if lifecycle_row is None and event_type in _RUNTIME_LEDGER_FILL_EVENTS:
-            if (
-                row_order_id in carry_in_execution_order_ids
-                or row_execution_id in carry_in_execution_ids
-                or (
-                    not carry_in_execution_order_ids
-                    and not carry_in_execution_ids
-                    and not carry_in_execution_symbols
-                    and _source_authority_order_event_row(row)
-                )
-            ):
-                lifecycle_row = _runtime_lifecycle_ledger_row(
-                    row,
-                    event_type=event_type,
-                )
-                if lifecycle_row is not None:
-                    lifecycle_row["source"] = "order_feed_lifecycle"
-        if lifecycle_row is not None:
-            carry_in_ledger_rows.append(lifecycle_row)
-    combined_order_lifecycle_rows = [
-        *(carry_in_order_lifecycle_rows or []),
-        *(order_lifecycle_rows or []),
-    ]
-    combined_event_sourced_fill_economics_order_ids = (
-        _event_sourced_fill_economics_order_ids(combined_order_lifecycle_rows)
-    )
-    for row in carry_in_execution_rows or []:
-        ledger_fill, _ledger_event_times = _runtime_execution_ledger_fill_from_row(
-            row,
-            order_lifecycle_rows=combined_order_lifecycle_rows,
-            event_sourced_fill_economics_order_ids=(
-                combined_event_sourced_fill_economics_order_ids
-            ),
-        )
-        if ledger_fill is not None:
-            carry_in_ledger_rows.append(ledger_fill)
-    if not event_times:
-        return []
-    unique_times = sorted(set(event_times))
-    (
-        carry_in_ledger_rows,
-        active_carry_in_order_ids,
-        active_carry_in_decision_ids,
-    ) = _active_carry_in_ledger_rows(
-        carry_in_ledger_rows,
-        bucket_start=unique_times[0],
-    )
-    carry_in_execution_rows = _filter_carry_in_source_rows_for_active_lots(
-        carry_in_execution_rows,
-        active_order_ids=active_carry_in_order_ids,
-        active_decision_ids=active_carry_in_decision_ids,
-    )
-    carry_in_decision_lifecycle_rows = _filter_carry_in_source_rows_for_active_lots(
-        carry_in_decision_lifecycle_rows,
-        active_order_ids=active_carry_in_order_ids,
-        active_decision_ids=active_carry_in_decision_ids,
-    )
-    carry_in_order_lifecycle_rows = _filter_carry_in_source_rows_for_active_lots(
-        carry_in_order_lifecycle_rows,
-        active_order_ids=active_carry_in_order_ids,
-        active_decision_ids=active_carry_in_decision_ids,
-    )
-    bucket_ranges = [(unique_times[0], unique_times[-1] + timedelta(microseconds=1))]
-    realized_rows: list[dict[str, object]] = []
-    for bucket in build_runtime_ledger_buckets(
-        ledger_rows,
-        bucket_ranges=bucket_ranges,
-        group_by=("symbol",),
-        require_order_lifecycle=True,
-        carry_in_rows=carry_in_ledger_rows,
-    ):
-        if bucket.closed_trade_count <= 0 and not bucket.blockers:
-            continue
-        row = _runtime_ledger_tca_row_from_bucket(
-            bucket=bucket,
-            computed_at=unique_times[-1],
-        )
-        bucket_payload = row.get("runtime_ledger_bucket")
-        source_context = _runtime_source_context_for_bucket(
-            bucket=bucket,
-            execution_rows=execution_rows,
-            decision_lifecycle_rows=decision_lifecycle_rows,
-            order_lifecycle_rows=order_lifecycle_rows,
-            unlinked_order_lifecycle_rows=unlinked_order_lifecycle_rows,
-            carry_in_execution_rows=carry_in_execution_rows,
-            carry_in_decision_lifecycle_rows=carry_in_decision_lifecycle_rows,
-            carry_in_order_lifecycle_rows=carry_in_order_lifecycle_rows,
-        )
-        source_account_labels = cast(list[str], source_context["source_account_labels"])
-        source_decision_mode_counts = cast(
-            dict[str, int], source_context["source_decision_mode_counts"]
-        )
-        source_decision_profit_proof_eligible = bool(
-            source_context["source_decision_profit_proof_eligible"]
-        )
-        target_notional_sizing = cast(
-            dict[str, object], source_context["paper_route_target_notional_sizing"]
-        )
-        target_notional_sizing_blockers = _metadata_text_list(
-            target_notional_sizing.get("blockers")
-        )
-        source_refs = cast(list[str], source_context["source_refs"])
-        source_row_counts = cast(dict[str, int], source_context["source_row_counts"])
-        trade_decision_ids = cast(list[str], source_context["trade_decision_ids"])
-        execution_ids = cast(list[str], source_context["execution_ids"])
-        execution_tca_metric_ids = cast(
-            list[str], source_context["execution_tca_metric_ids"]
-        )
-        execution_order_event_ids = cast(
-            list[str], source_context["execution_order_event_ids"]
-        )
-        source_window_ids = cast(list[str], source_context["source_window_ids"])
-        source_window_status_counts = cast(
-            dict[str, int], source_context["source_window_status_counts"]
-        )
-        source_window_classification_counts = cast(
-            dict[str, int], source_context["source_window_classification_counts"]
-        )
-        source_window_gap_count = int(source_context["source_window_gap_count"] or 0)
-        source_window_gap_ranges = cast(
-            list[Mapping[str, object]], source_context["source_window_gap_ranges"]
-        )
-        source_window_start = cast(datetime, source_context["source_window_start"])
-        source_window_end = cast(datetime, source_context["source_window_end"])
-        source_offsets = cast(
-            list[Mapping[str, object]], source_context["source_offsets"]
-        )
-        source_materialization = cast(
-            str | None, source_context["source_materialization"]
-        )
-        authority_class = cast(str | None, source_context["authority_class"])
-        authority_reason = cast(str | None, source_context["authority_reason"])
-        order_feed_lifecycle_complete = bool(
-            source_context["order_feed_lifecycle_complete"]
-        )
-        fill_economics_complete = bool(source_context["fill_economics_complete"])
-        execution_tca_required = bool(source_context["execution_tca_required"])
-        execution_tca_complete = bool(source_context["execution_tca_complete"])
-        bucket_order_feed_fill_lifecycle_blockers = cast(
-            list[str], source_context["order_feed_fill_lifecycle_blockers"]
-        )
-        bucket_source_materialization_blockers = list(
-            bucket_order_feed_fill_lifecycle_blockers
-        )
-        if (
-            execution_tca_required
-            and not execution_tca_complete
-            and EXECUTION_TCA_MISSING_BLOCKER
-            not in bucket_source_materialization_blockers
-        ):
-            bucket_source_materialization_blockers.append(EXECUTION_TCA_MISSING_BLOCKER)
-        if (
-            execution_tca_required
-            and not execution_tca_complete
-            and RUNTIME_LEDGER_EXECUTION_TCA_REFS_MISSING_BLOCKER
-            not in bucket_source_materialization_blockers
-        ):
-            bucket_source_materialization_blockers.append(
-                RUNTIME_LEDGER_EXECUTION_TCA_REFS_MISSING_BLOCKER
-            )
-        equity_denominator = cast(
-            tuple[Decimal, str] | None, source_context["equity_denominator"]
-        )
-        if flat_start_authority is not None and isinstance(bucket_payload, Mapping):
-            row.update(flat_start_authority)
-            bucket_payload = {
-                **dict(bucket_payload),
-                **flat_start_authority,
-            }
-            row["runtime_ledger_bucket"] = bucket_payload
-        if source_account_labels and isinstance(bucket_payload, Mapping):
-            bucket_payload = {
-                **dict(bucket_payload),
-                "source_account_labels": source_account_labels,
-            }
-            if len(source_account_labels) == 1:
-                bucket_payload["source_account_label"] = source_account_labels[0]
-            row["runtime_ledger_bucket"] = bucket_payload
-        if source_decision_mode_counts:
-            row["source_decision_mode_counts"] = source_decision_mode_counts
-            row["profit_proof_eligible"] = source_decision_profit_proof_eligible
-            if isinstance(bucket_payload, Mapping):
-                source_decision_mode = next(
-                    iter(source_decision_mode_counts),
-                    None,
-                )
-                bucket_payload = {
-                    **dict(bucket_payload),
-                    "source_decision_mode_counts": source_decision_mode_counts,
-                    "profit_proof_eligible": source_decision_profit_proof_eligible,
-                }
-                if len(source_decision_mode_counts) == 1:
-                    row["source_decision_mode"] = source_decision_mode
-                    bucket_payload["source_decision_mode"] = source_decision_mode
-                if not source_decision_profit_proof_eligible:
-                    blockers = list(bucket_payload.get("blockers") or [])
-                    if (
-                        SOURCE_DECISION_MODE_NOT_PROFIT_PROOF_ELIGIBLE_BLOCKER
-                        not in blockers
-                    ):
-                        blockers.append(
-                            SOURCE_DECISION_MODE_NOT_PROFIT_PROOF_ELIGIBLE_BLOCKER
-                        )
-                    bucket_payload["blockers"] = blockers
-                    row["runtime_ledger_blockers"] = blockers
-                    row["post_cost_promotion_eligible"] = False
-                    row["promotion_blocker"] = (
-                        SOURCE_DECISION_MODE_NOT_PROFIT_PROOF_ELIGIBLE_BLOCKER
-                    )
-                row["runtime_ledger_bucket"] = bucket_payload
-        if target_notional_sizing.get("requires_target_notional_sizing"):
-            row["paper_route_target_notional_sizing"] = target_notional_sizing
-            if isinstance(bucket_payload, Mapping):
-                bucket_payload = {
-                    **dict(bucket_payload),
-                    "paper_route_target_notional_sizing": target_notional_sizing,
-                }
-                if target_notional_sizing_blockers:
-                    blockers = list(bucket_payload.get("blockers") or [])
-                    for blocker in target_notional_sizing_blockers:
-                        if blocker not in blockers:
-                            blockers.append(blocker)
-                    bucket_payload["blockers"] = blockers
-                    row["runtime_ledger_blockers"] = blockers
-                    row["post_cost_promotion_eligible"] = False
-                    row["promotion_blocker"] = target_notional_sizing_blockers[0]
-                row["runtime_ledger_bucket"] = bucket_payload
-        if isinstance(bucket_payload, Mapping):
-            bucket_payload = _with_runtime_ledger_source_authority_context(
-                bucket_payload,
-                source_window_start=source_window_start,
-                source_window_end=source_window_end,
-                source_refs=source_refs,
-                source_row_counts=source_row_counts,
-                trade_decision_ids=trade_decision_ids,
-                execution_ids=execution_ids,
-                execution_tca_metric_ids=execution_tca_metric_ids,
-                execution_order_event_ids=execution_order_event_ids,
-                source_window_ids=source_window_ids,
-                source_offsets=source_offsets,
-                source_window_status_counts=source_window_status_counts,
-                source_window_classification_counts=source_window_classification_counts,
-                source_window_gap_count=source_window_gap_count,
-                source_window_gap_ranges=source_window_gap_ranges,
-                order_feed_lifecycle_complete=order_feed_lifecycle_complete,
-                execution_economics_complete=fill_economics_complete,
-                execution_tca_required=execution_tca_required,
-                source_materialization=source_materialization,
-                authority_class=authority_class,
-                authority_reason=authority_reason,
-            )
-            row["runtime_ledger_bucket"] = bucket_payload
-            row["runtime_ledger_blockers"] = _metadata_text_list(
-                bucket_payload.get("blockers")
-            )
-        source_backed_runtime_ledger = (
-            allow_authoritative_runtime_ledger_materialization
-            and fill_economics_complete
-            and not bucket_source_materialization_blockers
-            and isinstance(bucket_payload, Mapping)
-            and not runtime_ledger_promotion_source_authority_blockers(bucket_payload)
-        )
-        event_sourced_runtime_ledger = (
-            source_backed_runtime_ledger
-            and isinstance(bucket_payload, Mapping)
-            and _runtime_ledger_bucket_profit_proof_present(bucket_payload)
-        )
-        if event_sourced_runtime_ledger:
-            row["authoritative"] = True
-            row["authority_reason"] = "event_sourced_runtime_ledger_profit_proof"
-            row["pnl_derivation"] = "execution_order_events_runtime_ledger"
-            row["post_cost_promotion_eligible"] = True
-            row["runtime_ledger_blockers"] = []
-            row.pop("promotion_blocker", None)
-            if isinstance(bucket_payload, Mapping):
-                runtime_bucket = {
-                    **dict(bucket_payload),
-                    "authoritative": True,
-                    "authority_reason": "event_sourced_runtime_ledger_profit_proof",
-                    "pnl_derivation": "execution_order_events_runtime_ledger",
-                    "source_materialization": (
-                        source_materialization or "source_execution_lifecycle"
-                    ),
-                }
-                if equity_denominator is not None:
-                    runtime_bucket["account_equity"] = str(equity_denominator[0])
-                    runtime_bucket["account_equity_source"] = equity_denominator[1]
-                row["runtime_ledger_bucket"] = runtime_bucket
-        elif source_backed_runtime_ledger:
-            row["post_cost_expectancy_basis"] = POST_COST_BASIS_RUNTIME_LEDGER
-            row["post_cost_promotion_eligible"] = False
-            row["authoritative"] = False
-            row["authority_reason"] = (
-                "source_execution_lifecycle_materialized_runtime_ledger"
-            )
-            row["pnl_derivation"] = "execution_order_events_runtime_ledger"
-            blockers = list(row.get("runtime_ledger_blockers") or [])
-            row["runtime_ledger_blockers"] = blockers
-            if blockers:
-                row["promotion_blocker"] = blockers[0]
-            else:
-                row.pop("promotion_blocker", None)
-            if isinstance(bucket_payload, Mapping):
-                runtime_bucket = {
-                    **dict(bucket_payload),
-                    "blockers": blockers,
-                    "pnl_basis": POST_COST_BASIS_RUNTIME_LEDGER,
-                    "authoritative": False,
-                    "authority_reason": (
-                        "source_execution_lifecycle_materialized_runtime_ledger"
-                    ),
-                    "pnl_derivation": "execution_order_events_runtime_ledger",
-                    "source_materialization": (
-                        source_materialization or "source_execution_lifecycle"
-                    ),
-                }
-                if equity_denominator is not None:
-                    runtime_bucket["account_equity"] = str(equity_denominator[0])
-                    runtime_bucket["account_equity_source"] = equity_denominator[1]
-                row["runtime_ledger_bucket"] = runtime_bucket
-        else:
-            row["post_cost_expectancy_basis"] = POST_COST_BASIS_EXECUTION_RECONSTRUCTION
-            row["post_cost_promotion_eligible"] = False
-            row["authoritative"] = False
-            row["authority_reason"] = (
-                "execution_reconstruction_not_runtime_ledger_proof"
-            )
-            row["pnl_derivation"] = "execution_reconstructed_from_execution_rows"
-            row["promotion_blocker"] = (
-                "execution_reconstruction_not_runtime_ledger_proof"
-            )
-            blockers = list(row.get("runtime_ledger_blockers") or [])
-            for blocker in bucket_source_materialization_blockers:
-                if blocker not in blockers:
-                    blockers.append(blocker)
-            if isinstance(bucket_payload, Mapping):
-                for blocker in runtime_ledger_promotion_source_authority_blockers(
-                    bucket_payload
-                ):
-                    if blocker not in blockers:
-                        blockers.append(blocker)
-            if "execution_reconstruction_not_runtime_ledger_proof" not in blockers:
-                blockers.append("execution_reconstruction_not_runtime_ledger_proof")
-            row["runtime_ledger_blockers"] = blockers
-            if isinstance(bucket_payload, Mapping):
-                bucket_payload = dict(bucket_payload)
-                bucket_payload["blockers"] = blockers
-                bucket_payload["pnl_basis"] = POST_COST_BASIS_EXECUTION_RECONSTRUCTION
-                bucket_payload["authoritative"] = False
-                bucket_payload["authority_reason"] = (
-                    "execution_reconstruction_not_runtime_ledger_proof"
-                )
-                if equity_denominator is not None:
-                    bucket_payload["account_equity"] = str(equity_denominator[0])
-                    bucket_payload["account_equity_source"] = equity_denominator[1]
-                row["runtime_ledger_bucket"] = bucket_payload
-        realized_rows.append(row)
-    return realized_rows
 
 
 def _load_json_artifact(ref: str) -> dict[str, Any]:
